Change structure - Apply suggestions
