+++ src/pages/be-square/components/_card.js
@@ -1,8 +1,14 @@
 import React, { useState } from 'react'
 import PropTypes from 'prop-types'
-import { Bullet, CardWrapper, DropdownWrapper, IconWrapper, ListContainer } from '../style/_card'
-import { TextWrapper } from '../style/_common'
-import { Minimize, Maximize } from '../images/_what-lies-ahead'
+import {
+    Bullet,
+    CardWrapper,
+    DropdownWrapper,
+    IconWrapper,
+    ListContainer,
+} from '../static/style/_card'
+import { TextWrapper } from '../static/style/_common'
+import { Minimize, Maximize } from '../static/images/_what-lies-ahead'
 
 const Card = ({ card_content, custom_icon, has_list, style, title_component }) => {
     const [is_list_open, setIsListOpen] = useState(false)
