+++ src/pages/be-square/index.js
@@ -1,13 +1,13 @@
 import React from 'react'
-import Advantages from './static/components/_advantages'
-import ApplicationProcess from './static/components/_application-process'
-import Description from './static/components/_description'
-import Hero from './static/components/_hero'
-import JoinUs from './static/components/_join-us'
-import WhatIsBeSquare from './static/components/_what-is_besquare'
-import WhatLiesAhead from './static/components/_what-lies-ahead'
-import WhoCanApply from './static/components/_who-can-apply'
-import WhyDeriv from './static/components/_why-deriv'
+import Advantages from './components/_advantages'
+import ApplicationProcess from './components/_application-process'
+import Description from './components/_description'
+import Hero from './components/_hero'
+import JoinUs from './components/_join-us'
+import WhatIsBeSquare from './components/_what-is_besquare'
+import WhatLiesAhead from './components/_what-lies-ahead'
+import WhoCanApply from './components/_who-can-apply'
+import WhyDeriv from './components/_why-deriv'
 import Layout from 'components/layout/layout'
 import { WithIntl } from 'components/localization'
